Less y-flip hacks, more fractional values

+++ Skeleton3/cpp/common/Data/Controller.h
@@ -269,7 +269,7 @@ public:
      * @param y the y-coordinate of the desired pixel.
      * @return the value of the pixel at (x, y), or blank if it could not be obtained.
      */
-    QString getPixelValue( int x, int y );
+    QString getPixelValue( double x, double y );
 
     virtual ~Controller();
